Ali/Sprint1/AliDemoRepo/requirements.txt
aws-cdk.core==1.135.0
aws-cdk.assertions==1.135.0
aws-cdk.aws.lambda==1.135.0
aws-cdk.aws_events_targets==1.135.0
aws-cdk.aws_events==1.135.0
aws-cdk.aws_iam==1.135.0
aws-cdk.aws_cloudwatch==1.135.0
aws-cdk.aws_sns==1.135.0
aws-cdk.aws_sns_subscriptions==1.135.0
aws-cdk.aws_cloudwatch_actions==1.135.0
aws-cdk.aws_dynamodb==1.135.0
aws-cdk.aws_sqs==1.135.0
aws-cdk.aws_s3==1.135.0
aws_cdk.pipelines==1.135.0
aws-cdk.aws_codepipeline_actions==1.135.0
